공지사항 관리자만 삭제버튼 볼수있게 설정

+++ GDJ47_ALLIN_semi/src/main/webapp/views/notice/noticeView.jsp
@@ -31,10 +31,13 @@
             <th>내 용</th>
             <td><%=n.getNoticeContent() %></td>
         </tr>
-        <tr>
+ 		<tr>
             <th colspan="2">
+            <%if(login!=null&&login.getUserId().equals("admin")){ %>
                 <button onclick="deleteNotice(<%=n.getNoticeNo() %>)">삭제</button>
+           <%} %>
             </th>
+        
         </tr>
     </table>
 </section>
